@@ -1,38 +1,27 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import styles from './clock.module.css';
 
-export default class Clock extends React.Component {
+function getCurrentTime() {
+	const date = new Date();
+	return date.toLocaleTimeString([], {hour: '2-digit', minute: '2-digit'});
+}
 
-	state = {
-		time: this.getCurrentTime(),
-		interval: null
-	}
+export default function Clock() {
+	const [time, setTime] = useState(getCurrentTime());
 
-	componentWillMount() {
+	useEffect(() => {
 		const interval = window.setInterval(() => {
-			this.setState({
-				time: this.getCurrentTime()
-			});
+			setTime(getCurrentTime());
 		}, 1000);
-		this.setState({interval});
-	}
 
-	componentWillUnmount() {
-		window.clearInterval(this.state.interval);
-	}
+		return () => window.clearInterval(interval);
+	}, []);
 
-	getCurrentTime() {
-		const date = new Date();
-		return date.toLocaleTimeString([], {hour: '2-digit', minute:'2-digit'});
-	}
-
-	render() {
-		return (
-			<div className={styles.clock}>
-				<div className="center-vertical">
-					{this.state.time}
-				</div>
+	return (
+		<div className={styles.clock}>
+			<div className="center-vertical">
+				{time}
 			</div>
-		);
-	}
+		</div>
+	);
 }
