@@ -3,8 +3,7 @@ module.exports = function(api) {
 
 	return {
 		plugins: [
-			'@babel/proposal-class-properties',
-			'@babel/proposal-object-rest-spread'
+			'@babel/proposal-class-properties'
 		],
 		presets: [
 			['@babel/env', {modules: false}],
