@@ -1,54 +1,47 @@
 import React from 'react';
-import {connect} from 'react-redux';
+import {useDispatch} from 'react-redux';
 
 import {launchApp} from 'src/redux/windows';
-import {Resume} from 'src/apps';
+import {Resume, Projects, MediaPlayer} from 'src/apps';
 
 import headshotSrc from 'static/img/about-me/headshot.png';
 import styles from './me.module.css';
 
-function Me(props) {
-	const myResume = (
+function AppLink({children, app}) {
+	const dispatch = useDispatch();
+
+	return (
 		<button
-			className={styles.openResume}
-			onClick={() => props.launchApp(Resume)}
+			className={styles.appLink}
+			onClick={() => dispatch(launchApp(app))}
 		>
-			my resume
+			{children}
 		</button>
 	);
+}
 
-	const musicILike = (
-		<a href="https://www.youtube.com/watch?v=eDzBpQ0GcoU" target="blank">
-			music I like
-		</a>
-	);
-
+export default function Me() {
 	return (
 		<>
 			<img
 				src={headshotSrc}
-				alt="photo of some nerd"
+				alt="photo of some nerd (me)"
 				className={styles.headshot}
 			/>
 			<p className={styles.paragraph}>
-				Thanks for checking out my website! My name is Christian Dinh and I'm
-				a software engineer living in Austin TX. Most of my experience is in front
-				end web dev, but I probably wouldn't mind the opportunity to branch out a bit.
-				If you're interested in a more detailed description of my skills, take a look at {myResume} :D
+				Thanks for checking out my website! My name is Christian Dinh and I'm a software engineer
+				living in NYC. Most of my experience is in front end web dev, but I probably wouldn't
+				mind the opportunity to branch out a bit. If you're interested in a more detailed description
+				of my experience take a look at <AppLink app={Resume}>my resume</AppLink>, and if you're
+				interested in examples of my work take a look at <AppLink app={Projects}>my projects</AppLink> :D
 			</p>
 			<p className={styles.paragraph}>
-				Outside of programming, I {'<3'} music! I play a few instruments (bass is my fav)
-				and go to a lot of live shows. Some of the {musicILike} is objectively unlistenable,
-				but we all have our flaws. Some other things I think are dope:
-				movies, Magic: The Gathering (mostly EDH and proxy vintage), video games, snowboarding,
-				cooking, cocktails, and long walks on the beach.
+				Outside of programming, I {'<3'} music! I play a few instruments (bass is my fav),
+				compose chiptune music on my gameboy, and go to a lot of live shows (when I'm not living
+				through a pandemic). Some of <AppLink app={MediaPlayer}>my music</AppLink> is hosted on
+				this website. Some other things I think are dope: movies, Magic: The Gathering (mostly EDH and
+				proxy vintage), video games, snowboarding, cooking, cocktails, and long walks on the beach.
 			</p>
 		</>
 	);
 }
-
-const mapDispatchToProps = {
-	launchApp
-};
-
-export default connect(null, mapDispatchToProps)(Me);
