@@ -1,23 +1,19 @@
-import {applyMiddleware, combineReducers, createStore} from 'redux';
-import {createLogger} from 'redux-logger';
+import {applyMiddleware, combineReducers, createStore, compose} from 'redux';
 import thunk from 'redux-thunk';
 import webBrowser from './web-browser';
 import windows from './windows';
+
+const devToolsCompose = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
 const reducer = combineReducers({
 	webBrowser,
 	windows
 });
 
-const logger = createLogger({
-	level: 'info',
-	timestamp: false
-});
-
 export default (defaultState = {}) => {
 	return createStore(
 		reducer,
 		defaultState,
-		applyMiddleware(logger, thunk)
+		devToolsCompose(applyMiddleware(thunk))
 	);
 };
